update imports and add common.ts
